+++ test/linearAlgebra/CSRMatrix.cpp
@@ -17,23 +17,23 @@ TEST_CASE("CSRMatrix")
     // sparse matrix
     NeoN::Vector<NeoN::scalar> valuesSparse(exec, {1.0, 5.0, 6.0, 8.0});
     NeoN::Vector<NeoN::localIdx> colIdxSparse(exec, {0, 1, 2, 1});
-    NeoN::Vector<NeoN::localIdx> rowPtrsSparse(exec, {0, 1, 3, 4});
+    NeoN::Vector<NeoN::localIdx> rowOffsSparse(exec, {0, 1, 3, 4});
     NeoN::la::CSRMatrix<NeoN::scalar, NeoN::localIdx> sparseMatrix(
-        valuesSparse, colIdxSparse, rowPtrsSparse
+        valuesSparse, colIdxSparse, rowOffsSparse
     );
     const NeoN::la::CSRMatrix<NeoN::scalar, NeoN::localIdx> sparseMatrixConst(
-        valuesSparse, colIdxSparse, rowPtrsSparse
+        valuesSparse, colIdxSparse, rowOffsSparse
     );
 
     // dense matrix
     NeoN::Vector<NeoN::scalar> valuesDense(exec, {1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0});
     NeoN::Vector<NeoN::localIdx> colIdxDense(exec, {0, 1, 2, 0, 1, 2, 0, 1, 2});
-    NeoN::Vector<NeoN::localIdx> rowPtrsDense(exec, {0, 3, 6, 9});
+    NeoN::Vector<NeoN::localIdx> rowOffsDense(exec, {0, 3, 6, 9});
     NeoN::la::CSRMatrix<NeoN::scalar, NeoN::localIdx> denseMatrix(
-        valuesDense, colIdxDense, rowPtrsDense
+        valuesDense, colIdxDense, rowOffsDense
     );
     const NeoN::la::CSRMatrix<NeoN::scalar, NeoN::localIdx> denseMatrixConst(
-        valuesDense, colIdxDense, rowPtrsDense
+        valuesDense, colIdxDense, rowOffsDense
     );
 
     // NOTE: The purpose of this test is to detect changes in the order
@@ -46,17 +46,17 @@ TEST_CASE("CSRMatrix")
         auto valuesDenseHostView = valuesDenseHost.view();
         auto colIdxDenseHost = colIdxDense.copyToHost();
         auto colIdxDenseHostView = colIdxDenseHost.view();
-        auto rowPtrsDenseHost = rowPtrsDense.copyToHost();
-        auto rowPtrsDenseHostView = rowPtrsDenseHost.view();
+        auto rowOffsDenseHost = rowOffsDense.copyToHost();
+        auto rowOffsDenseHostView = rowOffsDenseHost.view();
 
         for (int i = 0; i < valuesDenseHostView.size(); ++i)
         {
             REQUIRE(valuesDenseHostView[i] == values[i]);
             REQUIRE(colIdxDenseHostView[i] == colIdxs[i]);
         }
-        for (int i = 0; i < rowPtrsDenseHostView.size(); ++i)
+        for (int i = 0; i < rowOffsDenseHostView.size(); ++i)
         {
-            REQUIRE(rowPtrsDenseHostView[i] == rowOffs[i]);
+            REQUIRE(rowOffsDenseHostView[i] == rowOffs[i]);
         }
     }
 
@@ -280,11 +280,11 @@ TEST_CASE("CSRMatrix")
         auto [value, column, row] = hostMatrix.view();
         auto hostvaluesSparse = valuesSparse.copyToHost();
         auto hostcolIdxSparse = colIdxSparse.copyToHost();
-        auto hostrowPtrsSparse = rowPtrsSparse.copyToHost();
+        auto hostrowOffsSparse = rowOffsSparse.copyToHost();
 
         REQUIRE(hostvaluesSparse.size() == value.size());
         REQUIRE(hostcolIdxSparse.size() == column.size());
-        REQUIRE(hostrowPtrsSparse.size() == row.size());
+        REQUIRE(hostrowOffsSparse.size() == row.size());
 
         for (NeoN::localIdx i = 0; i < value.size(); ++i)
         {
@@ -293,7 +293,7 @@ TEST_CASE("CSRMatrix")
         }
         for (NeoN::localIdx i = 0; i < row.size(); ++i)
         {
-            REQUIRE(hostrowPtrsSparse.view()[i] == row[i]);
+            REQUIRE(hostrowOffsSparse.view()[i] == row[i]);
         }
     }
 }