+++ test/linearAlgebra/petsc.cpp
@@ -45,8 +45,8 @@ TEST_CASE("MatrixAssembly - Petsc")
         Vector<NeoN::scalar> values(exec, {10.0, 4.0, 7.0, 2.0, 10.0, 8.0, 3.0, 6.0, 10.0});
         // TODO work on support for unsingned types
         Vector<localIdx> colIdx(exec, {0, 1, 2, 0, 1, 2, 0, 1, 2});
-        Vector<localIdx> rowPtrs(exec, {0, 3, 6, 9});
-        CSRMatrix<scalar, localIdx> csrMatrix(values, colIdx, rowPtrs);
+        Vector<localIdx> rowOffs(exec, {0, 3, 6, 9});
+        CSRMatrix<scalar, localIdx> csrMatrix(values, colIdx, rowOffs);
 
         Vector<NeoN::scalar> rhs(exec, {1.0, 2.0, 3.0});
         LinearSystem<scalar, localIdx> linearSystem(csrMatrix, rhs);
@@ -77,8 +77,8 @@ TEST_CASE("MatrixAssembly - Petsc")
             );
 
             Vector<localIdx> colIdx(exec, {0, 1, 2, 0, 1, 2, 0, 1, 2});
-            Vector<localIdx> rowPtrs(exec, {0, 3, 6, 9});
-            CSRMatrix<scalar, localIdx> csrMatrix(values, colIdx, rowPtrs);
+            Vector<localIdx> rowOffs(exec, {0, 3, 6, 9});
+            CSRMatrix<scalar, localIdx> csrMatrix(values, colIdx, rowOffs);
 
             Vector<NeoN::scalar> rhs(exec, {1.0, 2.0, 3.0});
             LinearSystem<scalar, localIdx> linearSystem(csrMatrix, rhs);